UE4 - occlusion queries for foliage and grass
[CL 2468748 by Gil Gribb in Main branch]

+++ Engine/Source/Runtime/Engine/Classes/Components/PrimitiveComponent.h
@@ -426,7 +426,7 @@ protected:
 	TEnumAsByte<EHasCustomNavigableGeometry::Type> bHasCustomNavigableGeometry;
 
 	/** Next id to be used by a component. */
-	static uint64 NextComponentId;
+	static uint32 NextComponentId;
 
 public:
 	/**
